+++ HISTORY.md
@@ -70,6 +70,7 @@
 
 ### 2022-12-01
 * blocks: don't show "result pic" option in context menu with "click-to-run" disabled
+* blocks: added configuration option for "noRingify"
 
 ### 2022-11-30
 * German translation update